+++ src/components/mumbleCard.tsx
@@ -1,5 +1,5 @@
 import React, { FC, useReducer } from 'react';
-import { likePost, Mumble } from '../services/mumble';
+import { Mumble } from '../services/mumble';
 import {
   Card,
   CommentButton,
@@ -9,6 +9,7 @@ import {
 } from '@smartive-education/design-system-component-library-hello-world-team';
 import { useSession } from 'next-auth/react';
 import { CommentMumble } from './comment';
+import { likePost } from '../services/likes';
 
 interface MumbleCard {
   mumble: Mumble;
@@ -78,7 +79,7 @@ export const MumbleCard: FC<MumbleCard> = ({ mumble }) => {
           />
           <CopyButton onClick={undefined} active={false} label={{ inactive: 'Copy Link', active: 'Link copied' }} />
         </div>
-        <div>{state.showComment ? <CommentMumble mumbleId={mumble.id}></CommentMumble> : null}</div>
+        {state.showComment && <CommentMumble mumbleId={mumble.id}></CommentMumble>}
       </Card>
     </>
   );
