+++ src/services/posts.ts
@@ -1,4 +1,4 @@
-import { UploadImage, transformMumble } from './serviceTypes';
+import { transformMumble, UploadImage } from './serviceTypes';
 import axios from 'axios';
 
 export const postMumble = async (text: string, file: UploadImage | null, accessToken?: string) => {
@@ -23,4 +23,22 @@ export const postMumble = async (text: string, file: UploadImage | null, accessT
   } catch (error) {
     throw new Error(error instanceof Error ? error.message : 'Could not post mumble');
   }
+};
+
+export const commentPost = async (params: { postId: string; comment: string; accessToken?: string }) => {
+  const { postId, comment, accessToken } = params || {};
+
+  if (!accessToken) {
+    throw new Error('No access token');
+  }
+
+  return await axios.post(
+    `${process.env.NEXT_PUBLIC_QWACKER_API_URL}posts/${postId}`,
+    { text: comment },
+    {
+      headers: {
+        Authorization: `Bearer ${accessToken}`,
+      },
+    }
+  );
 };
